Delete all files

+++ openhumans/views.py
@@ -26,11 +26,20 @@ class delete_file(View):
         return redirect(self.not_authorized_url)
 
 
-def delete_all_oh_files(oh_member):
-    """
-    Delete all current project files in Open Humans for this project member.
-    """
-    ohapi.api.delete_files(
-        project_member_id=oh_member.oh_id,
-        access_token=oh_member.get_access_token(),
-        all_files=True)
+class delete_all_oh_files(View):
+    success_url = 'list'
+    not_authorized_url = 'index'
+
+    def get(self, request):
+        """
+        Delete all current project files in Open Humans for this project member.
+        """
+
+        if request.user.is_authenticated and request.user.username != 'admin':
+            oh_member = request.user.openhumansmember
+            ohapi.api.delete_files(
+                project_member_id=oh_member.oh_id,
+                access_token=oh_member.get_access_token(),
+                all_files=True)
+            return redirect(self.success_url)
+        return redirect(self.not_authorized_url)
